feat: create new files for ipv4 and domain enrichments
